#14 refacto and first collab creation when database is empty
